python-yolo/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
ultralytics>=8.1.0
Pillow>=10.2.0
python-multipart>=0.0.9
numpy>=1.26.0
opencv-python-headless>=4.9.0
